+++ Source/Core/DolphinWX/FifoPlayerDlg.h
@@ -56,7 +56,7 @@ private:
 	wxString CreateRecordingFifoSizeLabel() const;
 	wxString CreateRecordingMemSizeLabel() const;
 	wxString CreateRecordingFrameCountLabel() const;
-	wxString CreateIntegerLabel(int size, const wxString& label) const;
+	wxString CreateIntegerLabel(size_t size, const wxString& label) const;
 
 	bool GetSaveButtonEnabled() const;
